Deprecate public Go module and make it internal
